employee_management/requirements.txt
Django>=5.1.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
